Error handling - Heartbeat
Custom error handling. Created an enum to handle our different error
types in order to make it so that we can just return something of that
enum type whenever we encounter an error.

Also created a custom NoSuchVenue type, as Stockfighter's heartbeat
api is slightly misleading: it'll return the JSON encoded version of the
struct that you expect ... unless you're requesting the heartbeat of a
venue that doesn't exist (or was recently deallocated) in which case
a completely different struct is returned.